+++ CHANGELOG.rst
@@ -8,6 +8,27 @@ This project follows the guidelines of `Keep a changelog`_ and adheres to
 .. _Keep a changelog: http://keepachangelog.com/
 .. _Semantic versioning: https://semver.org/
 
+
+`0.1.5`_ - 2021-07-26
+=====================
+
+Fixed
+-----
+* Unable to retrieve quantity from units with fixed representation.
+
+
+`0.1.4`_ - 2021-07-22
+=====================
+
+Fixed
+-----
+* Dividing equal units yields the unit 'Hz' instead of '1'.
+
+Added
+-----
+* Representation for the power of units.
+
+
 `0.1.3`_ - 2021-03-08
 =====================
 
@@ -45,6 +66,8 @@ Added
 * Initial implementation.
 
 
+.. _0.1.5: https://github.com/emtpb/united/releases/tag/0.1.5
+.. _0.1.4: https://github.com/emtpb/united/releases/tag/0.1.4
 .. _0.1.3: https://github.com/emtpb/united/releases/tag/0.1.3
 .. _0.1.2: https://github.com/emtpb/united/releases/tag/0.1.2
 .. _0.1.1: https://github.com/emtpb/united/releases/tag/0.1.1